Fix hardware interrupts

+++ kernel/arch/i386/timer.c
@@ -16,14 +16,14 @@ timer_callback(registers_t* regs)
 }
 
 void
-init_timer(uint32_t frequency)
+init_timer(int16_t frequency)
 {
     register_irq_handler(IRQ0, &timer_callback);
 
     // The value we send to the PIT is the value to divide it's input clock
     // (1193180 Hz) by, to get our required frequency. Important to note is
     // that the divisor must be small enough to fit into 16-bits.
-    uint32_t divisor = 1193180 / frequency;
+    int16_t divisor = 1193180 / frequency;
 
     // send command byte.
     outb(0x43, 0x36);
